+++ src/AutoTable.jsx
@@ -1,7 +1,9 @@
 import React from "react";
+import { polyfill } from "react-lifecycles-compat";
 import PropTypes from "prop-types";
 import { Data } from "perch-data";
 import { TableRow, TableCell, Typography } from "material-ui";
+import isEqual from "lodash.isequal";
 import { BaseTable as Table, LoadingRow } from "./";
 
 const ErrorRow = ({ columnCount }) => (
@@ -18,26 +20,41 @@ ErrorRow.propTypes = {
   columnCount: PropTypes.number.isRequired
 };
 
+const getSortFromOrdering = ordering => {
+  let sortColumn = null;
+  let sortDirection = null;
+
+  if (ordering) {
+    const isFirstCharDirection = ordering.slice(0, 1) === "-";
+    sortColumn = isFirstCharDirection ? ordering.slice(1) : ordering;
+    sortDirection = isFirstCharDirection ? "desc" : "asc";
+  }
+
+  return { sortColumn, sortDirection };
+};
+
 class AutoTable extends React.Component {
+  static getDerivedStateFromProps(nextProps, prevState) {
+    if (!isEqual(nextProps.filter, prevState.filter)) {
+      const { sortColumn, sortDirection } = getSortFromOrdering(
+        nextProps.initialOrdering
+      );
+      return { filter: nextProps.filter, page: 1, sortColumn, sortDirection };
+    }
+    return null;
+  }
+
   constructor(props) {
     super(props);
 
-    const { initialOrdering = null } = props;
-    let sortColumn = null;
-    let sortDirection = null;
-
-    if (initialOrdering) {
-      const firstCharIsDirection = initialOrdering.slice(0, 1) === "-";
-      sortColumn = firstCharIsDirection
-        ? initialOrdering.slice(1)
-        : initialOrdering;
-      sortDirection = firstCharIsDirection ? "desc" : "asc";
-    }
+    const { filter, initialOrdering, rowsPerPage } = props;
+    const { sortColumn, sortDirection } = getSortFromOrdering(initialOrdering);
 
     this.state = {
+      filter,
       ordering: initialOrdering,
       page: 1,
-      rowsPerPage: props.rowsPerPage,
+      rowsPerPage,
       search: null,
       sortColumn,
       sortDirection
@@ -82,7 +99,7 @@ class AutoTable extends React.Component {
     this.setState({ ordering, sortColumn: column, sortDirection: direction });
   };
 
-  handleSearch = search => this.setState({ search });
+  handleSearch = search => this.setState({ search, page: 1 }); // Reset page on search
 
   handleChangePage = page => this.setState({ page });
 
@@ -92,14 +109,14 @@ class AutoTable extends React.Component {
     const {
       action,
       columns,
-      filter,
       headerPadding,
-      paginatable,
+      pageable,
       searchable,
       sortable
     } = this.props;
 
     const {
+      filter,
       ordering,
       page,
       rowsPerPage,
@@ -123,7 +140,7 @@ class AutoTable extends React.Component {
             onSearch={this.handleSearch}
             onSort={this.handleSort}
             pagination={
-              paginatable ? this.getPaginationForData(result.data) : null
+              pageable ? this.getPaginationForData(result.data) : null
             }
             searchable={searchable}
             sortColumn={sortColumn}
@@ -152,7 +169,7 @@ AutoTable.propTypes = {
   filter: PropTypes.object, // eslint-disable-line react/forbid-prop-types
   headerPadding: PropTypes.string,
   initialOrdering: PropTypes.string,
-  paginatable: PropTypes.bool,
+  pageable: PropTypes.bool,
   rowsPerPage: PropTypes.number,
   rowsPerPageOptions: PropTypes.arrayOf(PropTypes.number),
   searchable: PropTypes.bool,
@@ -162,12 +179,14 @@ AutoTable.propTypes = {
 AutoTable.defaultProps = {
   filter: null,
   headerPadding: "default",
-  paginatable: false,
+  pageable: false,
   rowsPerPage: null,
   rowsPerPageOptions: [10, 25, 50, 100],
   searchable: false,
   initialOrdering: null,
   sortable: false
 };
+
+polyfill(AutoTable);
 
 export default AutoTable;